Added new module 'intervals' and tests.
